requirements.txt
# FastAPI & Web Server
fastapi==0.109.0
uvicorn[standard]==0.27.0
python-multipart==0.0.6

# Document Processing
python-docx==1.1.0
diff-match-patch==20230430

# Azure SDKs
azure-storage-blob==12.19.0
azure-identity==1.15.0
azure-keyvault-secrets==4.7.0
openai==1.12.0

# Database
sqlalchemy==2.0.25
asyncpg==0.29.0
psycopg2-binary==2.9.9

# Redis & Celery
celery==5.3.6
redis==5.0.1

# AI/LLM
langchain==0.1.6
langchain-openai==0.0.5
langgraph==0.0.20
pydantic==2.6.0
pydantic-settings==2.1.0

# Observability
opentelemetry-api==1.22.0
opentelemetry-sdk==1.22.0
opentelemetry-instrumentation-fastapi==0.43b0
azure-monitor-opentelemetry==1.2.0

# Utilities
python-dotenv==1.0.0
colorama==0.4.6

